requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docopt==0.6.2
# mysqlclient==2.0.2
Pillow==8.1.2
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pyflakes==2.2.0
Pygments==2.7.3
PyJWT==2.0.0
pytz==2020.4
pyvim==3.0.2
six==1.15.0
sqlparse==0.4.1
wcwidth==0.2.5